+++ lib/src/db.rs
@@ -95,6 +95,7 @@ impl Db {
     }
 
     /// Search for a value, get a PropSubjectMap. If it does not exist, create a new one.
+    #[tracing::instrument(skip(self))]
     pub fn get_prop_subject_map(&self, string_val: &str) -> AtomicResult<PropSubjectMap> {
         let prop_sub_map = self
             .index_vals
@@ -279,12 +280,14 @@ impl Storelike for Db {
         }
     }
 
+    #[tracing::instrument(skip(self))]
     fn get_resource_extended(
         &self,
         subject: &str,
         skip_dynamic: bool,
         for_agent: Option<&str>,
     ) -> AtomicResult<Resource> {
+        tracing::trace!("get_resource_extended: {}", subject);
         // This might add a trailing slash
         let mut url = url::Url::parse(subject)?;
         let clone = url.clone();
@@ -430,6 +433,7 @@ impl Storelike for Db {
     }
 
     // TPF implementation that used the index_value cache, far more performant than the StoreLike implementation
+    #[tracing::instrument(skip(self))]
     fn tpf(
         &self,
         q_subject: Option<&str>,
@@ -438,6 +442,7 @@ impl Storelike for Db {
         // Whether resources from outside the store should be searched through
         include_external: bool,
     ) -> AtomicResult<Vec<Atom>> {
+        tracing::trace!("tpf");
         let mut vec: Vec<Atom> = Vec::new();
 
         let hassub = q_subject.is_some();