+++ server/src/handlers/commit.rs
@@ -34,6 +34,6 @@ pub async fn post_commit(
         .commit_monitor
         .do_send(crate::actor_messages::CommitMessage { commit_response });
 
-    log::info!("{}", &message);
+    tracing::info!("{}", &message);
     Ok(builder.body(message))
 }