+++ server/src/handlers/resource.rs
@@ -53,7 +53,8 @@ pub async fn get_resource(
     timer.add("get_agent");
 
     let mut builder = HttpResponse::Ok();
-    log::info!("get_resource: {} as {}", subject, content_type.to_mime());
+
+    tracing::info!("get_resource: {} as {}", subject, content_type.to_mime());
     builder.header("Content-Type", content_type.to_mime());
     // This prevents the browser from displaying the JSON response upon re-opening a closed tab
     // https://github.com/joepio/atomic-data-rust/issues/137