+++ CHANGELOG.md
@@ -5,6 +5,9 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 
 ## [unreleased]
 
+- Replace `log` with `tracing` for structured logging #261
+- Add `--log-level` option
+- Add `--trace-chrome` option
 - Correct 404 status code
 - Server-Timings header #256
 - Fix indented welcome message